v1.3.1, UI/UX update

+++ app/Http/Controllers/SettingsController.php
@@ -27,27 +27,31 @@ class SettingsController extends Controller
     }
 
     public function updateSkin(){
-        list($width, $height) = getimagesize(Input::file('skin'));
-        if($height > 128 || $width > 128){
-            Session::flash('bad', 'Plik skórki jest za duży, musi on mieć przynajmniej rozmiar 64x32 i być w formacie PNG.');
+        if(!empty(Input::file('skin'))){
+            list($width, $height) = getimagesize(Input::file('skin'));
+            if($height > 128 || $width > 128){
+                Session::flash('bad', 'Plik skórki jest za duży, musi on mieć przynajmniej rozmiar 64x32 i być w formacie PNG.');
+            }else{
+                $skinid = bin2hex(openssl_random_pseudo_bytes(16));
+                $user = Skin::find(Auth::user()->id);
+                if($user->skin !== "0000000000000000000000000000000f") {
+                    Storage::delete('public/skins/' . $user->skin);
+                }
+                $user->skin = $skinid;
+
+                if(Input::get('skin_type') === '1')
+                    $user->skin_type = 'slim';
+                else{
+                    $user->skin_type = 'steve';
+                }
+
+                $user->save();
+                Storage::putFileAs('public/skins/', Input::file('skin'),$skinid);
+                Session::flash('success', 'Pomyślnie zmieniono skórkę!');
+            };
         }else{
-            $skinid = bin2hex(openssl_random_pseudo_bytes(16));
-            $user = Skin::find(Auth::user()->id);
-            if($user->skin !== "0000000000000000000000000000000f") {
-                Storage::delete('public/skins/' . $user->skin);
-            }
-            $user->skin = $skinid;
-
-            if(Input::get('skin_type') === '1')
-                $user->skin_type = 'slim';
-            else{
-                $user->skin_type = 'steve';
-            }
-
-            $user->save();
-            Storage::putFileAs('public/skins/', Input::file('skin'),$skinid);
-            Session::flash('success', 'Pomyślnie zmieniono skórkę!');
-        };
+            Session::flash('bad', 'Niewskazano pliku skórki, wskaż plik i pamiętaj, że musi on mieć przynajmniej rozmiar 64x32 i być w formacie PNG.');
+        }
         return view('user.skinchange');
     }
